+++ src/main/java/radon/jujutsu_kaisen/block/domain/SkyBlock.java
@@ -14,8 +14,8 @@ import radon.jujutsu_kaisen.block.JJKBlocks;
 import radon.jujutsu_kaisen.block.entity.DomainBlockEntity;
 import radon.jujutsu_kaisen.block.entity.JJKBlockEntities;
 
-public class UnlimitedVoidBlock extends DomainBlock implements EntityBlock {
-    public UnlimitedVoidBlock(Properties pProperties) {
+public class SkyBlock extends DomainBlock implements EntityBlock {
+    public SkyBlock(Properties pProperties) {
         super(pProperties);
     }
 
@@ -26,11 +26,11 @@ public class UnlimitedVoidBlock extends DomainBlock implements EntityBlock {
 
     @Override
     public @Nullable BlockEntity newBlockEntity(@NotNull BlockPos pPos, @NotNull BlockState pState) {
-        return JJKBlockEntities.UNLIMITED_VOID.get().create(pPos, pState);
+        return JJKBlockEntities.SKY.get().create(pPos, pState);
     }
 
     @Override
     public @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(@NotNull Level pLevel, @NotNull BlockState pState, @NotNull BlockEntityType<T> pBlockEntityType) {
-        return pLevel.isClientSide ? null : JJKBlocks.createTickerHelper(pBlockEntityType, JJKBlockEntities.UNLIMITED_VOID.get(), DomainBlockEntity::tick);
+        return pLevel.isClientSide ? null : JJKBlocks.createTickerHelper(pBlockEntityType, JJKBlockEntities.SKY.get(), DomainBlockEntity::tick);
     }
 }