+++ src/main/java/radon/jujutsu_kaisen/client/JJKShaders.java
@@ -10,14 +10,22 @@ import java.io.IOException;
 
 public class JJKShaders {
     private static ShaderInstance unlimitedVoidShader;
+    private static ShaderInstance skyShader;
 
     public static void onRegisterShaders(RegisterShadersEvent event) throws IOException {
         event.registerShader(new ShaderInstance(event.getResourceProvider(),
                         new ResourceLocation(JujutsuKaisen.MOD_ID, "unlimited_void"), DefaultVertexFormat.POSITION),
                 shader -> unlimitedVoidShader = shader);
+        event.registerShader(new ShaderInstance(event.getResourceProvider(),
+                        new ResourceLocation(JujutsuKaisen.MOD_ID, "sky"), DefaultVertexFormat.POSITION),
+                shader -> skyShader = shader);
     }
 
     public static ShaderInstance getUnlimitedVoidShader() {
         return unlimitedVoidShader;
+    }
+
+    public static ShaderInstance getSkyShader() {
+        return skyShader;
     }
 }